requirements.txt
rich
speedtest-cli
python-dotenv
